% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blankmemes}
\alias{blankmemes}
\title{All the memes available for meme creation in the memer package}
\format{
A data frame with 37 rows and 2 variables:
\describe{
\item{filename}{File where the image is}
\item{name}{Name to use to call the meme in \code{\link[=meme_get]{meme_get()}}}
}
}
\source{
\url{https://knowyourmeme.com/}
}
\usage{
blankmemes
}
\description{
All the memes available for meme creation in the memer package
}
\keyword{datasets}
