% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-meme.R
\name{get_meme}
\alias{get_meme}
\title{Read in the meme you wish to create}
\usage{
get_meme(memename)
}
\description{
Read in the meme you wish to create
}
\section{To add a new meme to the data}{

\describe{
   \item{First} Add the meme to the data folder (png)
   \item{Then} Add row to the blankmemes data \code{dplyr::add_row(filename = "...", name = "...")}
   \item{Finally} Run \code{usethis::use_data(blankmemes, overwrite = T)}
}
}

\examples{
get_meme("AllTheThings")
}
