% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-meme.R
\name{meme_add}
\alias{meme_add}
\title{Add meme to \code{memer} package}
\usage{
meme_add(name, file)
}
\arguments{
\item{name}{Name of the meme in CamelCase.}

\item{file}{Filename including the file extension.}
}
\description{
Add meme to \code{memer} package
}
\details{
\code{memer} uses \code{memer::blankmemes} to resolve filenames, so to use a
new meme you must build and install the package after running \code{meme_add()}.
}
\examples{
\dontrun{
memer:::meme_add("YoDawg", "yodawg.png")
}
}
\keyword{internal}
