% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-meme.R
\name{meme_explain}
\alias{meme_explain}
\title{Explain a meme (for the old people out there)}
\usage{
meme_explain(memename, launch = interactive(), content_warning = TRUE)
}
\arguments{
\item{memename}{A character describing the meme to get. See \code{\link[=meme_list]{meme_list()}}}

\item{launch}{Whether to launch the link in your browser
(defaults to TRUE if in interactive session)}

\item{content_warning}{Whether to warn about potentially offensive content
on knowyourmeme}
}
\value{
The link to the knowyourmeme article about a meme
}
\description{
Takes a meme name from memer and takes you to the
knowyourmeme page that can explain it for us old
people.
}
\examples{
meme_explain("Hotlinedrake", launch = FALSE)
}
