% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-meme.R
\name{meme_get}
\alias{meme_get}
\title{Read in the meme you wish to create}
\usage{
meme_get(memename)
}
\arguments{
\item{memename}{A character describing the meme to get. See \code{\link{meme_list}}}
}
\description{
Read in the meme you wish to create
}
\section{To add a new meme to the data}{

\describe{
   \item{First}{Add the meme to the data folder (png)}
   \item{Then}{Add row to the blankmemes data \code{dplyr::add_row(filename = "...", name = "...")}}
   \item{Finally}{Run \code{usethis::use_data(blankmemes, overwrite = T)}}
}
}

\examples{
meme_get("AllTheThings")
}
