% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{meme_text_top}
\alias{meme_text_top}
\alias{meme_text_bottom}
\title{Place text in the meme}
\usage{
meme_text_top(img, txt, font = "Impact", color = "white",
  strokecolor = "black", gravity = "North", size = 44, ...)

meme_text_bottom(img, txt, font = "Impact", color = "white",
  strokecolor = "black", gravity = "South", size = 44, ...)
}
\arguments{
\item{img}{An image of class \code{magick_image}. See \code{\link{meme_get}}.}

\item{txt}{A character string}

\item{font}{string with font family such as `"sans"`, `"mono"`, `"serif"`, `"Times"`, `"Helvetica"`, `"Trebuchet"`, `"Georgia"`, `"Palatino"` or `"Comic Sans"`.}

\item{color}{a valid [color string](https://www.imagemagick.org/Magick++/Color.html) such as `"navyblue"` or `"#000080"`}

\item{strokecolor}{a [color string](https://www.imagemagick.org/Magick++/Color.html) adds a stroke (border around the text)}

\item{gravity}{string with [gravity](https://www.imagemagick.org/Magick++/Enumerations.html#GravityType) value from [gravity_types].}

\item{size}{font-size in pixels}

\item{...}{Arguments passed on to \code{magick::image_annotate}
\describe{
  \item{location}{geometry string with location relative to \code{gravity}}
  \item{degrees}{rotates text around center point}
  \item{boxcolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
for background color that annotation text is rendered on.}
}}
}
\description{
Place text in the meme
}
\section{Functions}{
\itemize{
\item \code{meme_text_top}: Place text in center top of meme

\item \code{meme_text_bottom}: Place text in bottom center of meme
}}

\examples{
meme_get("AllTheThings") \%>\% meme_text_top("GGPLOT ALL THE THINGS!")
meme_get("OprahGiveaway") \%>\% meme_text_bottom("EVERYONE GETS A \%>\%!!!")
}
