% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-meme.R
\name{meme_tweet}
\alias{meme_tweet}
\title{Posts a meme to Twitter}
\usage{
meme_tweet(meme,
  tweet_text = "It's dangerous to meme alone. Here take this.",
  tag_rstatsmemes = FALSE, token = NULL)
}
\arguments{
\item{meme}{an external pointer of class 'magick-image'}

\item{tweet_text}{The text that you want to post along with your meme}

\item{tag_rstatsmemes}{adds two hashtags and tags the "rstatsmemes" account}

\item{token}{Every user should have their own Oauth (Twitter API) token. By
default \code{token = NULL} this function looks for the path to a saved
Twitter token via environment variables (which is what \code{create_token()}
sets up by default during initial token creation). For instruction on how
to create a Twitter token see the tokens vignette, i.e.,
\code{vignettes("auth", "rtweet")} or see \code{?tokens}.}
}
\description{
Make sure to have rtweet installed!
}
\examples{
\dontrun{
meme_get("OprahGiveaway") \%>\% 
  meme_text_bottom("EVERYONE GETS R MEMES!!!", size = 36) \%>\% 
  meme_tweet(tweet_text = "It's dangerous to meme alone. Here take this.", 
  tag_rstatsmemes = T)
}
}
