% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memer.R
\docType{package}
\name{memer}
\alias{memer}
\title{\code{memer} package}
\description{
Wrapper functions around \code{magick} package functions to create memes in R.
}
\details{
See the README on
or \href{https://github.com/sctyner/memer#README}{GitHub}
}
