% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{top_text}
\alias{top_text}
\alias{bottom_text}
\title{Place text in the meme}
\usage{
top_text(img, txt, ...)

bottom_text(img, txt, ...)
}
\arguments{
\item{img}{An image of class \code{magick_image}. See \code{\link[=get_meme]{get_meme()}}.}

\item{txt}{A character string}

\item{...}{Arguments passed on to \code{magick::image_annotate}
\describe{
  \item{color}{a valid \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}} such as
\code{"navyblue"} or \code{"#000080"}}
  \item{gravity}{string with
\href{{https://www.imagemagick.org/Magick++/Enumerations.html#GravityType}{gravity}}
value from \link{gravity_types}.}
  \item{location}{geometry string with location relative to \code{gravity}}
  \item{degrees}{rotates text around center point}
  \item{size}{font-size in pixels}
  \item{font}{string with font family such as \code{"sans"}, \code{"mono"}, \code{"serif"},
\code{"Times"}, \code{"Helvetica"}, \code{"Trebuchet"}, \code{"Georgia"}, \code{"Palatino"} or \code{"Comic Sans"}.}
  \item{strokecolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
adds a stroke (border around the text)}
  \item{boxcolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
for background color that annotation text is rendered on.}
}}
}
\description{
Place text in the meme
}
\section{Functions}{
\itemize{
\item \code{top_text}: Place text in center top of meme

\item \code{bottom_text}: Place text in bottom center of meme
}}

\examples{
get_meme("AllTheThings") \%>\% top_text("R ALL THE THINGS!")
get_meme("OprahGiveaway") \%>\% bottom_text("EVERYONE GETS A \%>\%!!!")
}
