% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-meme.R
\name{tweet_meme}
\alias{tweet_meme}
\title{Posts a meme to Twitter}
\usage{
tweet_meme(meme,
  tweet_text = "It's dangerous to meme alone. Here take this.",
  tag_rstatsmemes = F)
}
\arguments{
\item{meme}{an external pointer of class 'magick-image'}

\item{tweet_text}{The text that you want to post along with your meme}

\item{tag_rstatsmemes}{adds two hashtags and tags the "rstatsmemes" account}
}
\description{
Posts a meme to Twitter
}
\examples{
meme_get("OprahGiveaway") \%>\% 
  meme_text_bottom("EVERYONE GETS R MEMES!!!", size = 36) \%>\% 
  tweet_meme(tweet_text = "It's dangerous to meme alone. Here take this.", tag_rstatsmemes = T)
}
