% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{meme_text_distbf}
\alias{meme_text_distbf}
\alias{meme_text_rock}
\alias{meme_text_batman}
\alias{meme_text_trump}
\alias{meme_text_pigeon}
\alias{meme_text_chopper}
\alias{meme_text_brain}
\alias{meme_text_kermit}
\alias{meme_text_buttons}
\alias{meme_text_drake}
\alias{meme_text_printer}
\alias{meme_text_anakin}
\title{Specialized functions for placing text in memes}
\usage{
meme_text_distbf(
  img,
  newgirl,
  guy,
  oldgirl,
  font = "Impact",
  color = "white",
  strokecolor = "black",
  size = 44,
  ...
)

meme_text_rock(
  img,
  dwayne,
  girl,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 16,
  ...
)

meme_text_batman(
  img,
  robin,
  batman,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 20,
  ...
)

meme_text_trump(
  img,
  rightpg,
  leftpg = "",
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 20,
  ...
)

meme_text_pigeon(
  img,
  isthis,
  humanoid,
  butterfly,
  font = "Impact",
  color = "white",
  strokecolor = "black",
  size = 30,
  ...
)

meme_text_chopper(
  img,
  sr1,
  jr1,
  sr2,
  jr2,
  sr3,
  font = "Impact",
  color = "white",
  strokecolor = "black",
  size = 20,
  ...
)

meme_text_brain(
  img,
  br1,
  br2,
  br3,
  br4,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 24,
  ...
)

meme_text_kermit(
  img,
  good,
  dark,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 20,
  ...
)

meme_text_buttons(
  img,
  lbtn,
  rbtn,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 18,
  ...
)

meme_text_drake(
  img,
  top,
  bot,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 30,
  ...
)

meme_text_printer(
  img,
  left,
  right,
  font = "Impact",
  color = "black",
  strokecolor = NULL,
  size = 13,
  ...
)

meme_text_anakin(
  img,
  anakin,
  padme,
  font = "Impact",
  size = 18,
  width = 35,
  ...
)
}
\arguments{
\item{img}{An image of class \code{magick_image}. See \code{\link[=meme_get]{meme_get()}}.}

\item{newgirl}{A character string. Text for the left-most person in the \code{"DistractedBf"} meme.}

\item{guy}{A character string. Text for the center person in the \code{"DistractedBf"} meme.}

\item{oldgirl}{A character string. Text for the right-most person in the \code{"DistractedBf"} meme.}

\item{font}{string with font family such as \code{"sans"}, \code{"mono"}, \code{"serif"}, \code{"Times"}, \code{"Helvetica"}, \code{"Trebuchet"}, \code{"Georgia"}, \code{"Palatino"} or \code{"Comic Sans"}.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string} such as \code{"navyblue"} or \code{"#000080"}}

\item{strokecolor}{a \href{https://www.imagemagick.org/Magick++/Color.html}{color string} adds a stroke (border around the text)}

\item{size}{font-size in pixels}

\item{...}{
  Arguments passed on to \code{\link[magick:painting]{magick::image_annotate}}
  \describe{
    \item{\code{location}}{geometry string with location relative to \code{gravity}}
    \item{\code{degrees}}{rotates text around center point}
    \item{\code{boxcolor}}{a \href{https://www.imagemagick.org/Magick++/Color.html}{color string}
for background color that annotation text is rendered on.}
  }}

\item{dwayne}{A character string. Text for the top speech bubble in the \code{"TheRockDriving"} meme.}

\item{girl}{A character string. Text for the other speech bubble in the \code{"TheRockDriving"} meme.}

\item{robin}{A character string. Text for the Robin's speech bubble in the \code{"BatmanRobin"} meme.}

\item{batman}{A character string. Text for the Batman's speech bubble in the \code{"BatmanRobin"} meme.}

\item{rightpg}{A character string. Text for the right page in the \code{"TrumpBillSigning"} meme.}

\item{leftpg}{A character string. Text for the left page in the \code{"TrumpBillSigning"} meme.}

\item{isthis}{A character string. Text for the question in the \code{"IsThisAPigeon"} meme.}

\item{humanoid}{A character string. Text for the humanoid in the \code{"IsThisAPigeon"} meme.}

\item{butterfly}{A character string. Text for the butterfly in the \code{"IsThisAPigeon"} meme.}

\item{sr1}{A character string. Text for the first panel in the \code{"AmericanChopper"} meme.}

\item{jr1}{A character string. Text for the second panel in the \code{"AmericanChopper"} meme.}

\item{sr2}{A character string. Text for the third panel in the \code{"AmericanChopper"} meme.}

\item{jr2}{A character string. Text for the fourth panel in the \code{"AmericanChopper"} meme.}

\item{sr3}{A character string. Text for the fifth panel in the \code{"AmericanChopper"} meme.}

\item{br1}{A character string. Text for the first panel in the \code{"ExpandingBrain"} meme.}

\item{br2}{A character string. Text for the second panel in the \code{"ExpandingBrain"} meme.}

\item{br3}{A character string. Text for the third panel in the \code{"ExpandingBrain"} meme.}

\item{br4}{A character string. Text for the fourth panel in the \code{"ExpandingBrain"} meme.}

\item{good}{A character string. A good idea for the \code{"OffRamp"} meme.}

\item{dark}{A character string. A tempting, bad idea for the \code{"OffRamp"} meme.}

\item{lbtn}{A character string. The left button in the \code{"TwoButtonsAnxiety"} meme.}

\item{rbtn}{A character string. The right button in the \code{"TwoButtonsAnxiety"} meme.}

\item{top}{A character string. Text for the top panel in the \code{"HotlineDrake"} meme.}

\item{bot}{A character string. Text for the bottom panel in the \code{"HotlineDrake"} meme.}

\item{left}{A character string. Text for the left panel in the \code{"MoneyPrinter"} meme.}

\item{right}{A character string. Text for the right panel in the \code{"MoneyPrinter"} meme.}

\item{anakin}{A character string. Text for Anakin  in the \code{"AnakinPadmeRight"} meme.}

\item{padme}{A character string. Text for Padme in the \code{"AnakinPadmeRight"} meme.}

\item{width}{positive integer giving target line width in characters. A
width less than or equal to 1 will put each word on its own line.}
}
\description{
Specialized functions for placing text in memes
}
\section{Functions}{
\itemize{
\item \code{meme_text_distbf}: Text function for the distracted boyfriend meme.

\item \code{meme_text_rock}: Text function for The Rock driving meme.

\item \code{meme_text_batman}: Text function for the Batman slaps Robin meme.

\item \code{meme_text_trump}: Text function for the Trump "first order of business" meme.

\item \code{meme_text_pigeon}: Text function for the Is this a pigeon? meme.

\item \code{meme_text_chopper}: Text function for the American Chopper Senior vs. Junior fight meme.

\item \code{meme_text_brain}: Text function for the expanding brain meme.

\item \code{meme_text_kermit}: Text function for the Off Ramp meme.

\item \code{meme_text_buttons}: Text function for the Two Buttons Anxiety meme.

\item \code{meme_text_drake}: Text function for the Drake meme.

\item \code{meme_text_printer}: Text function for the Money Printer meme.

\item \code{meme_text_anakin}: Text function for the Anakin/Padme meme.
}}

\examples{
meme_get("DistractedBf") \%>\%
  meme_text_distbf("tidyverse", "new R users", "base R")
meme_get("TheRockDriving") \%>\%
  meme_text_rock("What's your favorite thing to do in R?" , "Write for loops.")
meme_get("BatmanRobin") \%>\%
  meme_text_batman("Hey, I'm Batman!" , "No, you idiot, I'm Batman!")
meme_get("TrumpBillSigning") \%>\%
  meme_text_trump("Stuff and nonsense")
meme_get("IsThisAPigeon") \%>\%
  meme_text_pigeon("Is this a pigeon?" , "Me", "Not a pigeon")
meme_get("AmericanChopper") \%>\%
  meme_text_chopper("Stop calling yourself 'just an R user'!",
                    "But I've never written a package!",
                    "So?? You're still an important part of the R community!",
                    "But people who write packages are more important",
                    "NO! All members of the R community are valued & deserve respect!")
meme_get("ExpandingBrain") \%>\%
  meme_text_brain("text1",
                    "text2",
                    "text3",
                    "text4")
meme_get("EvilKermit") \%>\%
  meme_text_kermit("Me: Your text here", "Me: Your evil text here")

meme_get("TwoButtonsAnxiety") \%>\%
  meme_text_buttons("Should I do this?", "Or this?")

meme_get("HotlineDrake") \%>\%
  meme_text_drake("Handcrafted memes", "Reproducible memes")

meme_get("MoneyPrinter") \%>\%
  meme_text_printer("nooooo!!!!!! you can't just use open source software to
                    make memes without annoying watermarks on them nooooo",
                    "haha meme printer go brrrr")

meme_get("AnakinPadmeRight") \%>\%
  meme_text_anakin("I'm going to set my working directory",
                   "With the {here} package, right?")

}
