% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{meme_text_distbf}
\alias{meme_text_distbf}
\alias{meme_text_rock}
\alias{meme_text_batman}
\alias{meme_text_trump}
\alias{meme_text_pigeon}
\alias{meme_text_chopper}
\title{Specialized functions for placing text in memes}
\usage{
meme_text_distbf(img, newgirl, guy, oldgirl, ...)

meme_text_rock(img, dwayne, girl, ...)

meme_text_batman(img, robin, batman, ...)

meme_text_trump(img, rightpg, leftpg = "", ...)

meme_text_pigeon(img, txt, humanoid, butterfly, ...)

meme_text_chopper(img, sr1, jr1, sr2, jr2, sr3, ...)
}
\arguments{
\item{img}{An image of class \code{magick_image}. See \code{\link[=meme_get]{meme_get()}}.}

\item{newgirl}{A character string. Text for the left-most person in the \code{"DistractedBf"} meme.}

\item{guy}{A character string. Text for the center person in the \code{"DistractedBf"} meme.}

\item{oldgirl}{A character string. Text for the right-most person in the \code{"DistractedBf"} meme.}

\item{...}{Arguments passed on to \code{magick::image_annotate}
\describe{
  \item{color}{a valid \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}} such as
\code{"navyblue"} or \code{"#000080"}}
  \item{gravity}{string with
\href{{https://www.imagemagick.org/Magick++/Enumerations.html#GravityType}{gravity}}
value from \link{gravity_types}.}
  \item{location}{geometry string with location relative to \code{gravity}}
  \item{degrees}{rotates text around center point}
  \item{size}{font-size in pixels}
  \item{font}{string with font family such as \code{"sans"}, \code{"mono"}, \code{"serif"},
\code{"Times"}, \code{"Helvetica"}, \code{"Trebuchet"}, \code{"Georgia"}, \code{"Palatino"} or \code{"Comic Sans"}.}
  \item{strokecolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
adds a stroke (border around the text)}
  \item{boxcolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
for background color that annotation text is rendered on.}
}}

\item{dwayne}{A character string. Text for the top speech bubble in the \code{"therockdriving"} meme.}

\item{girl}{A character string. Text for the other speech bubble in the \code{"therockdriving"} meme.}

\item{robin}{A character string. Text for the Robin's speech bubble in the \code{"batmanrobin"} meme.}

\item{batman}{A character string. Text for the Batman's speech bubble in the \code{"batmanrobin"} meme.}

\item{rightpg}{A character string. Text for the right page in the \code{"trumpbill"} meme.}

\item{leftpg}{A character string. Text for the left page in the \code{"trumpbill"} meme.}

\item{humanoid}{A character string. Text for the humanoid in the \code{"IsThisAPigeon"} meme.}

\item{butterfly}{A character string. Text for the butterfly in the \code{"IsThisAPigeon"} meme.}

\item{sr1}{A character string. Text for the first panel.}

\item{jr1}{A character string. Text for the second panel.}

\item{sr2}{A character string. Text for the third panel.}

\item{jr2}{A character string. Text for the fourth panel.}

\item{sr3}{A character string. Text for the fifth panel.}
}
\description{
Specialized functions for placing text in memes
}
\section{Functions}{
\itemize{
\item \code{meme_text_distbf}: Text function for the distracted boyfriend meme.

\item \code{meme_text_rock}: Text function for The Rock driving meme.

\item \code{meme_text_batman}: Text function for the Batman slaps Robin meme.

\item \code{meme_text_trump}: Text function for the Trump "first order of business" meme.

\item \code{meme_text_pigeon}: Text function for the Is this a pigeon? meme.

\item \code{meme_text_chopper}: Text function for the American Chopper Senior vs. Junior fight meme.
}}

\examples{
meme_get("DistractedBf") \%>\%
  meme_text_distbf("tidyverse", "new R users", "base R")
meme_get("TheRockDriving") \%>\%
  meme_text_rock("What's your favorite thing to do in R?" , "Write for loops.")
meme_get("BatmanRobin") \%>\%
  meme_text_batman("Hey, I'm Batman!" , "No, you idiot, I'm Batman!")
meme_get("TrumpBillSigning") \%>\%
  meme_text_trump("Stuff and nonsense")
meme_get("IsThisAPigeon") \%>\%
  meme_text_pigeon("Is this a pigeon?" , "Me", "Not a pigeon")
meme_get("AmericanChopper") \%>\%
  meme_text_chopper("Stop calling yourself 'just an R user'!",
                    "But I've never written a package!",
                    "So?? You're still an important part of the community!",
                    "But people who write packages are more important",
                    "NO! All members of the community are valued & deserve respect!")
}
