% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{meme_text_distbf}
\alias{meme_text_distbf}
\alias{meme_text_rock}
\title{Specialized functions for placing text in memes}
\usage{
meme_text_distbf(img, newgirl, guy, oldgirl, ...)

meme_text_rock(img, dwayne, girl, ...)
}
\arguments{
\item{img}{An image of class \code{magick_image}. See \code{\link[=meme_get]{meme_get()}}.}

\item{newgirl}{A character string. Text for the left-most person in the \code{"distractedbf"} meme.}

\item{guy}{A character string. Text for the center person in the \code{"distractedbf"} meme.}

\item{oldgirl}{A character string. Text for the right-most person in the \code{"distractedbf"} meme.}

\item{...}{Arguments passed on to \code{magick::image_annotate}
\describe{
  \item{color}{a valid \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}} such as
\code{"navyblue"} or \code{"#000080"}}
  \item{gravity}{string with
\href{{https://www.imagemagick.org/Magick++/Enumerations.html#GravityType}{gravity}}
value from \link{gravity_types}.}
  \item{location}{geometry string with location relative to \code{gravity}}
  \item{degrees}{rotates text around center point}
  \item{size}{font-size in pixels}
  \item{font}{string with font family such as \code{"sans"}, \code{"mono"}, \code{"serif"},
\code{"Times"}, \code{"Helvetica"}, \code{"Trebuchet"}, \code{"Georgia"}, \code{"Palatino"} or \code{"Comic Sans"}.}
  \item{strokecolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
adds a stroke (border around the text)}
  \item{boxcolor}{a \href{{https://www.imagemagick.org/Magick++/Color.html}{color string}}
for background color that annotation text is rendered on.}
}}

\item{dwayne}{A character string. Text for the top speech bubble in the \code{"therockdriving"} meme.}

\item{girl}{A character string. Text for the other speech bubble in the \code{"therockdriving"} meme.}

\item{img}{An image of class \code{magick_image}. See \code{\link[=meme_get]{meme_get()}}.}
}
\description{
Specialized functions for placing text in memes
}
\section{Functions}{
\itemize{
\item \code{meme_text_distbf}: Text function for the distracted boyfriend meme.

\item \code{meme_text_rock}: Text function for The Rock driving meme.
}}

\examples{
meme_get("DistractedBf") \%>\%
  meme_text_distbf("tidyverse", "new R users", "base R")
meme_get("TheRockDriving") \%>\%
  meme_text_rock("What's your favorite thing to do in R?" , "Write for loops.")
}
