% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{meme_text_top}
\alias{meme_text_top}
\alias{meme_text_bottom}
\title{Place text in the meme}
\usage{
meme_text_top(
  img,
  txt,
  font = "Impact",
  color = "white",
  strokecolor = "black",
  gravity = "North",
  size = 44,
  ...
)

meme_text_bottom(
  img,
  txt,
  font = "Impact",
  color = "white",
  strokecolor = "black",
  gravity = "South",
  size = 44,
  ...
)
}
\arguments{
\item{img}{An image of class \code{magick_image}. See \code{\link[=meme_get]{meme_get()}}.}

\item{txt}{A character string}

\item{font}{string with font family such as \code{"sans"}, \code{"mono"}, \code{"serif"}, \code{"Times"}, \code{"Helvetica"}, \code{"Trebuchet"}, \code{"Georgia"}, \code{"Palatino"} or \code{"Comic Sans"}.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string} such as \code{"navyblue"} or \code{"#000080"}}

\item{strokecolor}{a \href{https://www.imagemagick.org/Magick++/Color.html}{color string} adds a stroke (border around the text)}

\item{gravity}{string with \href{https://www.imagemagick.org/Magick++/Enumerations.html#GravityType}{gravity} value from \link{gravity_types}.}

\item{size}{font-size in pixels}

\item{...}{
  Arguments passed on to \code{\link[magick:painting]{magick::image_annotate}}
  \describe{
    \item{\code{location}}{geometry string with location relative to \code{gravity}}
    \item{\code{degrees}}{rotates text around center point}
    \item{\code{boxcolor}}{a \href{https://www.imagemagick.org/Magick++/Color.html}{color string}
for background color that annotation text is rendered on.}
  }}
}
\description{
Place text in the meme
}
\section{Functions}{
\itemize{
\item \code{meme_text_top}: Place text in center top of meme

\item \code{meme_text_bottom}: Place text in bottom center of meme
}}

\examples{
meme_get("AllTheThings") \%>\% meme_text_top("GGPLOT ALL THE THINGS!")
meme_get("OprahGiveaway") \%>\% meme_text_bottom("EVERYONE GETS A \%>\%!!!")
}
\references{
Jeroen Ooms (2018). magick: Advanced Graphics and Image-Processing in R. R package version 2.0.
\url{https://CRAN.R-project.org/package=magick}
}
