% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-annotate.R
\name{top_text}
\alias{top_text}
\title{Place text on top of the meme.}
\usage{
top_text(img, txt, ...)
}
\description{
get_meme("AllTheThings") %>% top_text("R ALL THE THINGS!")
get_meme("OprahGiveaway") %>% bottom_text("EVERYONE GETS A %>%!!!")
get_meme("DistractedBf") %>% distbf_text("tidyverse", "new R users", "base R")
}
